# Conflicts:
#	epochStart/bootstrap/process.go
#	epochStart/errors.go
#	genesis/process/intermediate/standardDelegationProcessor.go
#	integrationTests/multiShard/smartContract/scCallingSC_test.go
#	process/block/preprocess/smartContractResults_test.go
#	process/block/preprocess/transactions_test.go
#	process/block/shardblock_test.go
#	process/coordinator/process_test.go
#	process/errors.go
#	process/interface.go
#	update/trigger/trigger_test.go
#	vm/systemSmartContracts/staking.go

+++ storage/txcache/eviction.go
@@ -51,7 +51,7 @@ func (cache *TxCache) isCapacityExceeded() bool {
 
 func (cache *TxCache) areThereTooManyBytes() bool {
 	numBytes := cache.NumBytes()
-	tooManyBytes := numBytes > uint64(cache.config.NumBytesThreshold)
+	tooManyBytes := numBytes > int(cache.config.NumBytesThreshold)
 	return tooManyBytes
 }
